retry_workers/requirements.txt
kafka-python
requests
pandas
psycopg2-binary